+++ test/scripts/run_linux_api_test_py_tim7xx.bash
@@ -1,12 +1,21 @@
 #!/bin/bash
 
+printf "\033c"
+
 # 
-# Run sick_generic_caller with sick_tim_7xx.launch (Linux native with test server)
+# Build and run minimalistic api usage examples (Python, C, C++)
+# 
+pushd ../../examples/scripts
+./build_run_api_examples_linux.bash
+popd
+
+# 
+# Run sick_scan_xd_api_test.py with sick_tim_7xx.launch (Linux native with test server)
 # 
 
-printf "\033c"
 pushd ../..
 export LD_LIBRARY_PATH=.:./build_linux:$LD_LIBRARY_PATH
+export PYTHONPATH=.:./python/api:$PYTHONPATH
 
 # Start emulator
 cp -f ./test/emulator/scandata/sopas_et_field_test_1_2_both_010.pcapng.json /tmp/lmd_scandata.pcapng.json